Refactored the application structure
The folder structure was made feature based.
