+++ src/common/SearchBox/SearchBox.spec.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, { shallow, mount } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import { SearchBox } from "../../src/components/SearchBox";
+import { SearchBox } from "./SearchBox";
 
 Enzyme.configure({
   adapter: new Adapter()
