+++ src/meal/MealPreview/MealPreview.spec.js
@@ -1,12 +1,7 @@
 import React from "react";
 import Enzyme, { mount } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import MealPreview, {
-  Thumb,
-  Title,
-  Info,
-  Button
-} from "../../src/widgets/MealPreview";
+import MealPreview, { Thumb, Title, Info, Button } from "./MealPreview";
 
 Enzyme.configure({
   adapter: new Adapter()
