--- src/pages/Browse.js
@@ -1,5 +1,0 @@
-import React, { Fragment } from "react";
-
-const Browse = () => <Fragment>Browse section</Fragment>;
-
-export default Browse;
