--- src/pages/NotFound.js
@@ -1,3 +1,0 @@
-const NotFound = () => "Page not found";
-
-export default NotFound;
